pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-lob-simulator"
version = "0.1.0"
description = "High-Frequency Limit Order Book Simulator for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.0",
    # Order book data structure (sorted price levels)
    "sortedcontainers>=2.4.0",
    # Numerical computations (volatility, returns, etc.)
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m env.server.app
server = "env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["env", "env.server"]
package-dir = { "env" = ".", "env.server" = "server" }
